aggregate function fixes

+++ src/AggregateFunctions/AggregateFunctionGroupBitmapData.h
@@ -37,7 +37,7 @@ private:
         rb = roaring_bitmap_create();
 
         for (const auto & x : small)
-            roaring_bitmap_add(rb, x.getValue());
+            roaring_bitmap_add(rb, static_cast<UInt32>(x.getValue()));
     }
 
 public:
@@ -62,12 +62,12 @@ public:
                 else
                 {
                     toLarge();
-                    roaring_bitmap_add(rb, value);
+                    roaring_bitmap_add(rb, static_cast<UInt32>(value)); // this does not seem right
                 }
             }
         }
         else
-            roaring_bitmap_add(rb, value);
+            roaring_bitmap_add(rb, static_cast<UInt32>(value)); // this does not seem right
     }
 
     UInt64 size() const
@@ -104,7 +104,7 @@ public:
             readStringBinary(s,in);
             rb = roaring_bitmap_portable_deserialize(s.c_str());
             for (const auto & x : small) // merge from small
-                roaring_bitmap_add(rb, x.getValue());
+                roaring_bitmap_add(rb, static_cast<UInt32>(x.getValue())); // this does not seem right
         }
         else
             small.read(in);
@@ -137,7 +137,7 @@ public:
     {
         roaring_bitmap_t * smallRb = roaring_bitmap_create();
         for (const auto & x : small)
-            roaring_bitmap_add(smallRb, x.getValue());
+            roaring_bitmap_add(smallRb, static_cast<UInt32>(x.getValue()));
         return smallRb;
     }
 
@@ -165,7 +165,7 @@ public:
         else if (isSmall() && r1.isLarge())
         {
             for (const auto & x : small)
-                if (roaring_bitmap_contains(r1.rb, x.getValue()))
+                if (roaring_bitmap_contains(r1.rb, static_cast<UInt32>(x.getValue())))
                     buffer.push_back(x.getValue());
 
             // Clear out the original values
